Add files via upload
Accepts the first output of the generator for autocorrelation studying. 
The utilized version of the receiver to test the generator output vector.

We can't use this program to actually receive the infra-sound signal because the first output bit wasn't readable as ASCII. The lazy but effective solution for this problem was to read but don't decode the first output bit.